# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Courses.API | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:5001/"
    }
  ],
  "paths": {
...
    "/api/tests/{testId}/attempts": {
      "post": {
        "tags": [
          "Tests"
        ],
        "parameters": [
          {
            "name": "testId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserTestAttemptResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemHttpResult"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Tests"
        ],
        "parameters": [
          {
            "name": "testId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserTestAttemptResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tests/attempts": {
      "get": {
        "tags": [
          "Tests"
        ],
        "parameters": [
          {
            "name": "PageIndex",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "OrderBy",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "OrderDirection",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedListOfUserTestAttemptResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/tests/attempts/{attemptId}": {
      "get": {
        "tags": [
          "Tests"
        ],
        "parameters": [
          {
            "name": "attemptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserTestAttemptResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemHttpResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/questions/{questionId}": {
      "get": {
        "tags": [
          "Questions"
        ],
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemHttpResult"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Questions"
        ],
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemHttpResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemHttpResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Questions"
        ],
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemHttpResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemHttpResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/questions": {
      "post": {
        "tags": [
          "Questions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemHttpResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnswerResponse": {
        "required": [
          "id",
          "value",
          "isCorrect"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "value": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          }
        }
      },
...
      "CreateAnswerDto": {
        "required": [
          "value",
          "isCorrect"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          }
        }
      },
      "CreateQuestionCommand": {
        "required": [
          "content",
          "order",
          "testId",
          "answers"
        ],
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "image": {
            "$ref": "#/components/schemas/IFormFile"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "testId": {
            "type": "string",
            "format": "uuid"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateAnswerDto"
            }
          }
        }
      },
...
      "IFormFile": {
        "type": "string",
        "format": "binary",
        "nullable": true
      },
...
      "PagedListOfUserTestAttemptResponse": {
        "required": [
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserTestAttemptResponse"
            }
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "nullable": true
      },
      "ProblemHttpResult": {
        "type": "object",
        "properties": {
          "problemDetails": {
            "$ref": "#/components/schemas/ProblemDetails"
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "QuestionResponse": {
        "required": [
          "id",
          "content",
          "order",
          "image",
          "testId",
          "answers"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "content": {
            "type": "string"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "testId": {
            "type": "string",
            "format": "uuid"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerResponse"
            }
          }
        }
      },
...
      "UpdateAnswerDto": {
        "required": [
          "value",
          "isCorrect"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          }
        }
      },
...
      "UpdateQuestionCommand": {
        "required": [
          "id",
          "content",
          "order",
          "image",
          "answers"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "content": {
            "type": "string"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateAnswerDto"
            }
          }
        }
      },
...
      "UserTestAnswerResponse": {
        "required": [
          "id",
          "questionId",
          "questionContent",
          "questionOrder",
          "possibleAnswers",
          "selectedAnswerIds"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "questionId": {
            "type": "string",
            "format": "uuid"
          },
          "questionContent": {
            "type": "string"
          },
          "questionOrder": {
            "type": "integer",
            "format": "int32"
          },
          "possibleAnswers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerResponse"
            }
          },
          "selectedAnswerIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "UserTestAttemptResponse": {
        "required": [
          "id",
          "userId",
          "userTestId",
          "isCompleted",
          "score",
          "createdAt",
          "completedAt",
          "answers"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "userTestId": {
            "type": "string",
            "format": "uuid"
          },
          "isCompleted": {
            "type": "boolean"
          },
          "score": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserTestAnswerResponse"
            }
          }
        }
      }
    }
  },
...
}